backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
databases[postgresql]
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
python-dotenv
Pillow
